Added admin privileges and code cleanup
